+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Birthdays: move today's birthdays out of notifications into right sidebar.
 - Performance: improve scrolling/navigation between feeds by using a virtual window library.
 - Admin: admin UI into 3-column layout.
+- Admin: include external link to frontend repo for the running commit.
 
 ### Removed
 - Theme: Halloween theme.